Implement comprehensive MCP export system
- Add MCP v1 schema models (Node, Edge, Pointer, Embedding, Manifest)
- Implement MCP export service with SAGE-to-Node mapping
- Add NDJSON writer for large collections
- Create manifest builder with checksums and metadata
- Add comprehensive validation and guardrails
- Build Dart CLI tool for MCP export
- Write extensive test suite with golden tests
- Add MCP documentation to README
- Support storage profiles (minimal, space_saver, balanced, hi_fidelity)
- Include privacy propagation and content-addressable storage
- Add deterministic export guarantees

Features:
- Standards-compliant MCP v1 export format
- SAGE narrative extraction from journal entries
- Content-addressable storage (CAS) URIs
- Privacy detection and propagation
- Deterministic exports with checksums
- Comprehensive validation system
- Command-line interface
- Programmatic API

Files created:
- lib/mcp/models/mcp_schemas.dart
- lib/mcp/export/mcp_export_service.dart
- lib/mcp/export/ndjson_writer.dart
- lib/mcp/export/manifest_builder.dart
- lib/mcp/export/checksum_utils.dart
- lib/mcp/validation/mcp_validator.dart
- tool/mcp/cli/arc_mcp_export.dart
- test/mcp_exporter_golden_test.dart

+++ ARC MVP/EPI/lib/media/analysis/audio_transcribe_service.dart
@@ -139,8 +139,8 @@ class WhisperStubTranscribeService implements AudioTranscribeService {
       final numSegments = (audioInfo.duration / windowSec).ceil();
       
       for (int i = 0; i < numSegments; i++) {
-        final startTime = i * windowSec.toDouble();
-        final endTime = ((i + 1) * windowSec.toDouble()).clamp(0, audioInfo.duration);
+        final startTime = (i * windowSec).toDouble();
+        final endTime = ((i + 1) * windowSec).toDouble().clamp(0.0, audioInfo.duration).toDouble();
         
         // Calculate segment data range
         final startByte = (startTime / audioInfo.duration * audioBytes.length).round();
@@ -251,7 +251,7 @@ class WhisperStubTranscribeService implements AudioTranscribeService {
         
         final transcript = _generateMockTranscript(endTime - startTime, enableDiarization);
         segments.add(TranscriptSegment(
-          ts: [startTime, endTime],
+          ts: [startTime.toDouble(), endTime.toDouble()],
           text: transcript,
         ));
       }
